+++ crates/zng-wgt-menu/src/lib.rs
@@ -26,6 +26,7 @@ use zng_wgt_input::{cursor, focus::alt_focus_scope, CursorIcon};
 use zng_wgt_size_offset::size;
 use zng_wgt_style::{impl_style_fn, style_fn, Style, StyleMix};
 use zng_wgt_text::Text;
+use zng_wgt_window::is_mobile;
 
 pub mod context;
 pub mod popup;
@@ -150,6 +151,7 @@ impl ButtonStyle {
                 });
                 align = Align::CENTER;
             };
+
             icon_fn = BUTTON.cmd().flat_map(|c| match c {
                 Some(c) => c.icon().boxed(),
                 None => LocalVar(WidgetFn::nil()).boxed()
@@ -164,6 +166,10 @@ impl ButtonStyle {
                 saturate = false;
                 opacity = 50.pct();
                 cursor = CursorIcon::NotAllowed;
+            }
+
+            when *#is_mobile {
+                shortcut_txt = NilUiNode;
             }
         }
     }