+++ crates/zng-wgt-menu/src/lib.rs
@@ -15,7 +15,7 @@ use colors::BASE_COLOR_VAR;
 use zng_ext_font::FontNames;
 use zng_ext_input::{focus::FOCUS, mouse::ClickMode};
 use zng_ext_l10n::lang;
-use zng_wgt::{align, base_color, is_disabled, margin, prelude::*};
+use zng_wgt::{align, base_color, is_disabled, is_mobile, margin, prelude::*};
 use zng_wgt_access::{access_role, AccessRole};
 use zng_wgt_button::BUTTON;
 use zng_wgt_container::{child_align, child_end, padding};
@@ -26,7 +26,6 @@ use zng_wgt_input::{cursor, focus::alt_focus_scope, CursorIcon};
 use zng_wgt_size_offset::size;
 use zng_wgt_style::{impl_style_fn, style_fn, Style, StyleMix};
 use zng_wgt_text::Text;
-use zng_wgt_window::is_mobile;
 
 pub mod context;
 pub mod popup;